% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilks.R
\name{wilks}
\alias{wilks}
\title{Calculate wilks points}
\usage{
wilks(body_weight, weight_lifted, gender, unit = "kg")
}
\arguments{
\item{body_weight}{Weight of the lifter}

\item{weight_lifted}{Total weight lifted}

\item{gender}{Either 'male' or 'female'}

\item{unit}{Either 'kg' or 'lbs'}
}
\value{
Wilks points
}
\description{
Calculate wilks points
}
\examples{
wilks(83, 450, 'male')
}
